# docs/openapi.yaml
openapi: 3.0.0
info:
  title: API Veterinaria
  description: 'Documentación para API Veterinaria'
  contact:
    email: franciscohanna92@gmail.com
  version: 1.0.0
servers:
- url: localhost
paths:
  /mascotas:
    get:
      tags:
      - Mascotas
      summary: Obtener un listado de las mascotas existentes
      operationId: obtenerMascotas
      responses:
        200:
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mascota'
        500:
          description: Hubo un error al obtener el listado de mascotas
          content: {}
    post:
      tags:
      - Mascotas
      summary: Agregar una nueva mascota
      operationId: agregarMascota
      requestBody:
        description: Objeto mascota a ser creado
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mascota'
        required: true
      responses:
        201:
          description: La mascota fue creada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mascota'
        400:
          description: Los datos de la mascota son incorrectos
          content: {}
        500:
          description: Ocurrió un error al crear la mascota
          content: {}
  /mascotas/{idMascota}:    
    delete:
      tags: 
        - Mascotas
      summary: Eliminar una mascota especifica
      operationId: eliminarMascota
      parameters: 
        - in: path
          name: idMascota
          required: true
          schema:
            type: string 
            example: "b11faa60-f287-11e9-8a31-dd8e324cb24b"
      responses:
        400:
          description: Operación Exitosa
          content: {}
        500:
          description: Hubo un error al eliminar la mascota
          content: {}  
    get:
      tags:
        - Mascotas
      summary: Buscar una mascota por su id
      operationId: buscarMascota
      parameters: 
      - in: path
        name: idMascota
        required: true
        schema:
          type: string
          example: "b11faa60-f287-11e9-8a31-dd8e324cb24b"
      responses:
        200:
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mascota'
        500:
          description: Hubo un error al eliminar la mascota
          content: {}
    patch:
      tags:
        - Mascotas
      summary: Modificar datos de una mascota por su id
      operationId: actualizarMascota
      parameters: 
      - in: path
        name: idMascota
        required: true
        schema:
          type: string
          example: "b11faa60-f287-11e9-8a31-dd8e324cb24b"
      responses:
        200:
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mascota'
        500:
          description: Hubo un error al modificar la mascota
          content: {}    
  /vacunas:    
    post:
      tags:
        - Vacunas
      summary: Agregar una nueva mascota
      operationId: agregarVacuna
      requestBody:
        description: Objeto vacuna a ser creado
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vacunas'
        required: true
      responses:
        201:
          description: La mascota fue creada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vacunas'
        400:
          description: Los datos de la mascota son incorrectos
          content: {}
        500:
          description: Ocurrió un error al crear la mascota
          content: {}        
components:
  schemas:
    Mascota:
      required:
      - nombre
      - tipo
      type: object
      properties:
        _id:
          type: integer
        nombre:
          type: string
          example: Pepito
        tipo:
          type: string
          enum: [PERRO, GATO]
          example: PERRO
        esDeRaza:
          type: boolean
          example: doggie
        fechaNacimiento:
          type: string
          format: date
          example: doggie
        # vacunas:
        #   type: array
        #   items:
        #     $ref: './vacuna.yaml'
    